+++ dashboard/app/controllers/print_certificates_controller.rb
@@ -1,6 +1,8 @@
 require 'base64'
 
 class PrintCertificatesController < ApplicationController
+  include CertificatesHelper
+
   # GET /print_certificates/:encoded_params
   def show
     prevent_caching
@@ -14,18 +16,5 @@ class PrintCertificatesController < ApplicationController
 
     @student_name = data['name']
     @certificate_image_url = certificate_image_url(data['name'], data['course'])
-  end
-
-  private
-
-  def certificate_image_url(name, course)
-    return '/images/hour_of_code_certificate.jpg' unless course
-
-    opts = {
-      name: name,
-      course: course
-    }.compact
-    encoded = Base64.urlsafe_encode64(opts.to_json)
-    "/certificate_images/#{encoded}.jpg"
   end
 end
